test: add huffman estimated gains
